api/requirements.txt
cffi==1.14.5
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==1.4.2
python-dotenv==0.18.0
six==1.16.0
uWSGI==2.0.19.1
Werkzeug==2.0.1
